chore(digital-data): moved declaration to _document.js
